# tools/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Pappaliiga Stats API",
                 "description":  "REST API for CS2 tournament statistics",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/matches/division/{championship_id}":  {
                                                                  "get":  {
                                                                              "tags":  [
                                                                                           "matches"
                                                                                       ],
                                                                              "summary":  "Get Division Matches",
                                                                              "description":  "Get all matches for a division.",
                                                                              "operationId":  "get_division_matches_api_matches_division__championship_id__get",
                                                                              "parameters":  [
                                                                                                 "@{name=championship_id; in=path; required=True; schema=}"
                                                                                             ],
                                                                              "responses":  {
                                                                                                "200":  "@{description=Successful Response; content=}",
                                                                                                "422":  "@{description=Validation Error; content=}"
                                                                                            }
                                                                          }
                                                              },
                  "/api/matches/{match_id}":  {
                                                  "get":  {
                                                              "tags":  [
                                                                           "matches"
                                                                       ],
                                                              "summary":  "Get Match Details",
                                                              "description":  "Get full match details including map results.",
                                                              "operationId":  "get_match_details_api_matches__match_id__get",
                                                              "parameters":  [
                                                                                 "@{name=match_id; in=path; required=True; schema=}"
                                                                             ],
                                                              "responses":  {
                                                                                "200":  "@{description=Successful Response; content=}",
                                                                                "422":  "@{description=Validation Error; content=}"
                                                                            }
                                                          }
                                              },
...
              },
...
}